requirements.txt
fastapi
uvicorn
python-jose[cryptography]
fastapi-mcp
openai
anthropic
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-multipart
ollama
textblob
language-tool-python
sqlalchemy
passlib[bcrypt]
python-multipart